+++ Paper/LayoutEngine/ChildElementInfo.cs
@@ -3,12 +3,12 @@
     internal class ChildElementInfo
     {
         public ElementHandle Element;
-        public double CrossBefore;
-        public double Cross;
-        public double CrossAfter;
-        public double MainBefore;
-        public double Main;
-        public double MainAfter;
+        public float CrossBefore;
+        public float Cross;
+        public float CrossAfter;
+        public float MainBefore;
+        public float Main;
+        public float MainAfter;
 
         public ChildElementInfo(ElementHandle element)
         {
