+++ Paper/LayoutEngine/StretchItem.cs
@@ -5,15 +5,15 @@
         public enum ItemTypes { Before, Size, After }
 
         public int Index { get; set; }
-        public double Factor { get; set; }
+        public float Factor { get; set; }
         public ItemTypes ItemType { get; set; }
-        public double Violation { get; set; } = 0f;
-        public double Computed { get; set; } = 0f;
+        public float Violation { get; set; } = 0f;
+        public float Computed { get; set; } = 0f;
         public bool Frozen { get; set; } = false;
-        public double Min { get; set; }
-        public double Max { get; set; }
+        public float Min { get; set; }
+        public float Max { get; set; }
 
-        public StretchItem(int index, double factor, ItemTypes itemType, double min, double max)
+        public StretchItem(int index, float factor, ItemTypes itemType, float min, float max)
         {
             Index = index;
             Factor = factor;
